+++ src/main/java/pl/pzsp2back/controllers/TradeOfferController.java
@@ -4,13 +4,20 @@ package pl.pzsp2back.controllers;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import pl.pzsp2back.dto.DtoMapper;
+import pl.pzsp2back.dto.TradeOfferDto;
 import pl.pzsp2back.dtoPost.TradeOfferPostDto;
 import pl.pzsp2back.exceptions.TradeOfferServiceException;
+import pl.pzsp2back.orm.TradeOffer;
+import pl.pzsp2back.orm.User;
 import pl.pzsp2back.services.TradeOfferService;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.stream.Collectors;
 
 @RestController
@@ -24,21 +31,31 @@ public class TradeOfferController {
     @Operation(summary = "Get trade offer with given id",
             description = "This endpoint returns trade offer for given id.")
     @GetMapping("/{id}")
-    public ResponseEntity<?> getTradeOffer(@PathVariable("id") Long id) {
+    public ResponseEntity<?> getTradeOffer(@AuthenticationPrincipal User requesterUser, @PathVariable("id") Long id) {
         try {
-            var offerDto = dtoMapper.toDto(tradeOfferService.getTradeOffer(id));
-            return ResponseEntity.ok(offerDto);
+            if ((requesterUser.getIfAdmin() && tradeOfferService.ifSameGroup(requesterUser.getLogin(), id)) || tradeOfferService.ifOfferOwner(requesterUser.getLogin(), id)) {
+                var offerDto = dtoMapper.toDto(tradeOfferService.getTradeOffer(id));
+                return ResponseEntity.ok(offerDto);
+            }
+            return new ResponseEntity<>("Action forbidden for this user.", HttpStatus.FORBIDDEN);
         } catch (TradeOfferServiceException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
+
+
     @Operation(summary = "Get all trade offers for given user (login)",
             description = "This endpoint returns all trade offers for given user, for all optimization processes, regardless if offer isActive or not, so you have to handle it.")
     @GetMapping("/all/{login}")
-    public ResponseEntity<?> getAllUserTradeOffers(@PathVariable("login") String login) {
+    public ResponseEntity<?> getAllUserTradeOffers(@AuthenticationPrincipal User requesterUser, @PathVariable("login") String login) {
         try {
-            var offerDtoList = tradeOfferService.getUserTradeOffers(login).stream().map(to -> dtoMapper.toDto(to)).collect(Collectors.toList());
+            List<TradeOfferDto> offerDtoList;
+            if(requesterUser.getIfAdmin()){
+                offerDtoList = tradeOfferService.getGroupTradeOffers(requesterUser.getLogin()).stream().map(to -> dtoMapper.toDto(to)).collect(Collectors.toList());
+            } else {
+                offerDtoList = tradeOfferService.getUserTradeOffers(requesterUser.getLogin()).stream().map(to -> dtoMapper.toDto(to)).collect(Collectors.toList());
+            }
             return ResponseEntity.ok(offerDtoList);
         } catch (TradeOfferServiceException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
@@ -55,10 +72,13 @@ public class TradeOfferController {
                     return: updated offer
                     """)
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateTradeOffer(@Valid @RequestBody TradeOfferPostDto tradeOfferPostDto, @PathVariable("id") Long id) {
+    public ResponseEntity<?> updateTradeOffer(@AuthenticationPrincipal User requesterUser, @Valid @RequestBody TradeOfferPostDto tradeOfferPostDto, @PathVariable("id") Long id) {
         try {
-            var updatedOfferDto = dtoMapper.toDto(tradeOfferService.updateTradeOffer(tradeOfferPostDto, id));
-            return ResponseEntity.ok(updatedOfferDto);
+            if (requesterUser.getIfAdmin() && tradeOfferService.ifSameGroup(requesterUser.getLogin(), id)) {
+                var updatedOfferDto = dtoMapper.toDto(tradeOfferService.updateTradeOffer(tradeOfferPostDto, id));
+                return ResponseEntity.ok(updatedOfferDto);
+            }
+            return new ResponseEntity<>("Action forbidden for this user.", HttpStatus.FORBIDDEN);
         } catch (TradeOfferServiceException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
@@ -71,15 +91,18 @@ public class TradeOfferController {
                     Price: can be null (default: 0)\n
                     TimeSlotId: must be given\n
                     IfWantOffer: must be given\n
-                    OptimizationProcessId: can be null (default: nearest optimization process - but not working yet...) \n
+                    OptimizationProcessId: can be null (default: nearest optimization process) \n
                     
                     return: updated offer
                     """)
     @PostMapping("/{login}")
-    public ResponseEntity<?> createTradeOffer(@Valid @RequestBody TradeOfferPostDto tradeOfferPostDto, @PathVariable("login") String login) {
+    public ResponseEntity<?> createTradeOffer(@AuthenticationPrincipal User requesterUser, @Valid @RequestBody TradeOfferPostDto tradeOfferPostDto, @PathVariable("login") String login) {
         try {
-            var createdOfferDto = dtoMapper.toDto(tradeOfferService.createTradeOffer(tradeOfferPostDto, login));
-            return ResponseEntity.ok(createdOfferDto);
+            if (!requesterUser.getIfAdmin()) {
+                var createdOfferDto = dtoMapper.toDto(tradeOfferService.createTradeOffer(tradeOfferPostDto, requesterUser.getLogin()));
+                return ResponseEntity.ok(createdOfferDto);
+            }
+            return new ResponseEntity<>("Action forbidden for this user.", HttpStatus.FORBIDDEN);
         } catch (TradeOfferServiceException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
@@ -91,10 +114,14 @@ public class TradeOfferController {
                     return: deleted offer
                     """)
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteTradeOffer(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteTradeOffer(@AuthenticationPrincipal User requesterUser, @PathVariable("id") Long id) {
         try {
-            var deletedOfferDto = dtoMapper.toDto(tradeOfferService.deleteTradeOffer(id));
-            return ResponseEntity.ok(deletedOfferDto);
+            if ((requesterUser.getIfAdmin() && tradeOfferService.ifSameGroup(requesterUser.getLogin(), id)) || tradeOfferService.ifOfferOwner(requesterUser.getLogin(), id)) {
+                var deletedOfferDto = dtoMapper.toDto(tradeOfferService.deleteTradeOffer(id));
+                return ResponseEntity.ok(deletedOfferDto);
+            }
+            return new ResponseEntity<>("Action forbidden for this user.", HttpStatus.FORBIDDEN);
+
         } catch (TradeOfferServiceException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
